Compare delays / uvws produced by CALC and PyEphem
Calculate the delay and (u,v,w) coordinates for a specified target and
KAT-7 antenna pair, using both CALC and PyEphem. In the process, figure
out the EOP mechanism provided by PyEphem via the delta_t function.
The delays are spot-on (0.2 picosecond error) but the uvw coordinates
are out by centimetres (or a 46 arcsec mispointing of baseline).