Improve SSN projection tests
Add corner cases to highlight funny (x, y) valid domain. Rework closure
test so that it does not depend on SIN anymore. Instead it now samples
directly from the valid domain which is a clipped unit circle. The rest
of the closure test follows directly from the corresponding SIN version.