Use print-like handler if logging is unconfigured
Instead of suppressing log events via a NullHandler and relying on
the (frequently unsuspecting) user to configure logging, do a very basicConfig-like configuration at the module level. This logs
warnings and above to sys.stderr using the basic format (not unlike
a plain old print statement), thereby ensuring that the user receives
these warnings and errors.

While this behaviour is strongly discouraged by the logging
documentation, the saving grace is that the configured handler
also has a filter that disables the handler once the user properly
configures logging at the application or script level.