Radec target accepts RA in decimal degrees
Right ascension is typically interpreted as hours when given as a
sexagesimal string (e.g. 12:34:56) but taken as degrees when given as
a decimal string (e.g. 12.34). For examples of this, see the Simbad,
VizieR and NED sites. Since katpoint passes the RA string directly to
ephem.hours(), a decimal value is currently interpreted as hours.
This change brings the behaviour in line with standard practice.