Use (az, el) coordinates for target separation
In the past, the angular separation between two targets was found
based on their apparent (ra, dec) coordinates, motivated by a cryptic
comment in the code that calls this "the most reliable common
coordinate frame in ephem". I cannot remember why this should be so.

Use good old (az, el) instead, as all targets have to provide this for
the telescope. The problem with apparent (ra, dec) is that azel targets
fakes it with astrometric (ra, dec), thereby introducing an error
when comparing any target with an azel target (a very common use case).