New Model class to serve as base for models
The Model class handles the construction of various models (pointing,
delay, flux density) from a set of parameters. Each parameter is
represented by a Parameter object. It has a name, a floating-point
value, units and doc strings, and methods that convert it to and from
a string representation.

The Model contains a sequence of Parameters and can be constructed
from and saved to a ConfigParser (aka ini) file or a description string.
The Model can also be displayed in compact and extended fashion.
Include unit tests (with 100% coverage, whoohoo!).