Make docstring writable and add Model.values()
Use a metaclass trick to make the Model class docstring (and those
of its subclasses) writable, which will allow the number of parameters
to be inserted into the docstring after the fact instead of being
hardcoded. This trick is not necessary for Python 3.3 and above.

Expose the parameter values as a list of floats via the dict-like
values() method.