Plumb delay model into Antenna object
This replaces the old ENU offset parameter. The position_enu Antenna
attribute is now extracted from the delay model. The boolean value of a
Model is the presence of any active parameters (i.e. a non-empty model).

A side effect is that an ENU offset with fewer than three components will
now be accepted (e.g. an offset with missing U component), as the Model
class can be constructed from any number of model parameters. Remove the
corresponding ValueError raise and adjust the tests accordingly. A future
version might make some model parameters mandatory, but this is considered
too complicated for now.