Flesh out CorrelatorDelays to do delay correction
The CorrelatorDelays class now calculates delay corrections as expected
by the correlator / data proxy. The calculations are done in this class
instead of the lower-level DelayModel class, which only serves as a
repository of parameter values. The advantage is that multiple antennas
can be handled efficiently in one go. The semantics of the delay model
is therefore split between DelayModel and CorrelatorDelays.

The calculations in CorrelatorDelays are based on the katproxy package
(specifically, the katproxy.proxy.db_proxy.DelayModel class and its
_calculate_tau_max and poco_delays methods), the katconf package
(the katconf.sysconf.AntennaConfig.geometric_delay_max method) and
katpoint.Target.geometric_delay. For now it assumes that the reference
antenna has the same lat-long-alt position as the reference part of
the main antennas (i.e. that the ENU offsets in the delay models apply
directly to the reference antenna).

In the process, extend DelayModel to be converted to and from a sequence
of parameter values in delay format instead of as path lengths. This is
the form of the delay model that is most useful for calculations, while
the path length version is more human friendly (and needed for pointing
too). This requires the EM wave speed associated with each delay
parameter. Make a simple assumption for the speed associated with the
fixed path lengths (a mix of cable and free space).

Remove parameter counter from docstring as this is an unnecessary
complication for delay models. Rename CAB_x to FIX_x as it is less
misleading (the fixed delays are only partly due to cables and also
include free air and electronics segments).