Calculate delay correction for multiple timestamps
Let DelayCorrection handle a sequence of timestamps, in line with the rest
of katpoint. This automatically triggers the calculation of slopes for
linear interpolation. The next_timestamp mechanism is still convenient for
real-time operation in the proxy, but otherwise looks like a bit of an
appendix since it represents the special case of a sequence of 2 timestamps.