Check antenna reference positions
For now, the reference antenna position must match the reference positions
of each antenna in the array (and its own reference position), so that the
ENU offset in each antenna's delay model directly represent the baseline
between that antenna and the reference antenna. Raise a ValueError if not
the case.