Add explicit default value to Parameter
Up to now the implicit default parameter value was 0.0. Make this explicit by
passing a default_value to Parameter on construction. This may be something
else in future, e.g. 1, or 1.22 for the implicit beam model in Antenna.
A parameter is active if its value is not equal to the default - expose this
as the __nonzero__ value of Parameter. Use default value during Model
construction and display. The new default value also enables parameters of
non-float type to be used in future.