Make Model.description more useful and compact
Before, the description string contained all the parameters in comma-separated
form. This was the format of pointing models on disk. On the other hand, the
Antenna object uses a shortened string form of the model that is
whitespace-separated and does not contain the final inactive parameters.

Now that there is an explicit model file format, it does not make sense to
support two different string formats. Make the description string into the
compact version so that it can plug directly into an Antenna.