Allow targets to be specified with any epoch.
When creating a target the epoch of the coordinates can be specified as a tag. 'B1900' or 'B1950' will use the inbuilt ephem dates for those epochs. Other epochs can be specified with a decimal string, specifying the Gregorian date of the epoch desired. These changes have been made for both 'radec' and 'gal' targets.

The construct_radec_target method now allows the user to specify the epoch of the coordinates as well, which can be in any format accepted by ephem.Date().

J2000 is the default epoch in all cases.