Make target epoch be specified using ephem defaults or 'Jxxxx.xx'.
The user can now specify an epoch using the besselian dates B1900, or B1950- which are both ephem defaults. Otherwise the user can specify an arbitrary epoch using the string Jxxxx.xx. An epoch specified in such a way is defined (using a convention that is typically obscure for astronomers) as the date corresponding to the difference in Julian years of the specified epoch from the Gregorian date of J2000.