More accurate string descriptions of objects
Parameters (and thus models) now default to repr() instead of str() to
serialise floats. This ensures that they can round-trip from serialised
form, and that __eq__ is exact. This is currently causing a test
failure, because converting delays from distance to time and back again
is not a round trip, and this is now being detected.

Antenna.description is still not exact, because it is defined to
represent angles as D:M:S and this does not round trip (particularly
with the accuracy provided by ephem). Instead, description_repr has been
added to give better accuracy. It still does not round trip precisely,
because it is necessary to convert radians to degrees and back again,
but it will be out by at most a few ulp, rather than metres.

Target.description still needs to be updated similarly.