Sanitise Unicode strings and bytes for ephem
The ephem.degrees and ephem.hours factories raise a cryptic TypeError if
a string input is of the wrong type. It expects non-Unicode strings on Python 2
and Unicode strings on Python 3. Normalise the input string to the appropriate
type, but also make sure that numbers are still accepted. This allows Unicode
strings for Target and Antenna descriptions as well as model parameter strings.

The methods will still fail with a UnicodeEncodeError on Python 2 or a
UnicodeDecodeError on Python 3 if the string contains non-ASCII characters.