Split out Target.uvw_basis function
This makes the linear transformation from ENU baselines to UVW
coordinates available to the user. This allows computing all UVW
coordinates for an array to be done almost O(n_ants) faster, because the
uvw_basis function only depends on the reference antenna and timestamps,
and not the second antenna in the baseline, and applying the transform
is very quick.