Fix sign of fringe phase and fringe rate
A positive delay is corrected by adding a negative delay and positive phase
for a positive centre frequency (the case of frequency ascending with channel).
This inadvertently broke since the code was copied from the KAT-7 DBE proxy,
which assumed frequencies descended with channel index and hence should have
had a negative centre frequency. The fringe phase correction should now have
opposite sign to the delay correction for the usual positive centre frequency.