Download AIPS sources instead of having copies
These are problematic due to an incompatible software license (GPL).
Rather download the relevant source files from the official repository
and patch them to work with f2py.

The original included files were based on the 31DEC08 AIPS release which
is now harder to obtain. Bite the bullet and use the latest versions of
DIRCOS and NEWPOS. This requires extra source file downloads and compiles.