Add description property for DelayCorrection object
This description isn't quite as complete as the other models, since
the DelayCorrection object contains many models.  Printing all of
their parameters as part of the description would be overly verbose.
E.g. the reference observer location would be repeated for each
antenna.  The compromise is to print the minimum information related
to the delays.  This helps the user verify that the correct delay
models for each antenna are in use.  It is printed over multiple lines
to make it more readable.  In addition, the option to instantiate a
DelayCorrection object from the text description is not implemented.