Allow delay centre to be offset from target
Since the `corrections` method deals with everything from the target
object to the final delays, it has to implement offsets and other
manipulations of the target coordinates too to get proper delays.
The alternative would have been to feed the final (az, el) tracks
directly into `corrections` (good for more complicated scans than
fixed offsets) but this leaves more work for the library user.

The offset parameters are passed in as their own dict (as opposed to
general kwargs) to make them more explicit. The offset info could also
have been set on the object instead of being passed from method to
method but this would have introduced unwanted state.