Correctly handle offsets in (ra, dec)
This is potentially more useful than offsets in (az, el) when it comes
to the delay centre. The `Target.plane_to_sphere` method switches to
the requested spherical coordinate system and its return values will
be (ra, dec) even though they are called `az` and `el` (leading to
some confusion even for the author). It is necessary to convert these
to (az, el) to fit in with the rest of the delay machinery.