Refactor to replace ants with ant_models
Only the name and delay model bits of an Antenna matter to the
DelayCorrection object. Strengthen this by storing a dict of delay
models instead of a list of antennas on the object. There is no more
need to fake antennas when constructing from a description string.
The JSON string now has ant_models as a dict since antenna ordering
is actually not that important here (the proxy will look up inputs by
name and not by index). Demote `inputs` to a convenient internal
attribute since the `ant_models` attribute is all that needs to be
user-facing (less repetition).