Allow extra delay to be fixed externally
The 'extra' delay added to all inputs ensures that all delay
corrections are positive as they should be in a causal system.
Previously this value was estimated as the maximum absolute delay
possible in the given array (still the default behaviour).

The problem is that this delay will differ from one subarray to the
next, based on the specific dishes involved. This is bad for pulsar
timing, who would like a stable phase centre with fixed and known
delays. While max_delay could have been calculated after the fact,
this makes it easier by providing a way for it to be fixed in e.g.
a config file for the whole array and any subdivision of it.