Add support for polarisation in flux density models
The extended model is a frequency-invariant Stokes vector by which the
frequency-dependent value is scaled. This is added as an extra four
parameters to the model.

Since katpoint ignores extraneous coefficients, it was previously legal
to pass values that were ignored but are now interpreted. To avoid more
problems with this in future, you now get a FutureWarning if you pass
too many parameters. To test this, unittest2 is now used in the unit
tests.

The unit test has been rearranged slightly to split up the test into the
different pieces of functionality.