Just accept that the pointing model has 22 params
Remove the elaborate scheme that inserts the number of model parameters
into the docstrings of the `PointingModel` class and `fit` method.
The number of parameters has never changed in 10 years, so hard-code it
instead. The elaborate scheme involved metaclasses and different code
on Python 2 and 3, for starters (mostly due to the `fit` class method).