Allow different target with same name in Catalogue
Relax the constraints on a Catalogue so that it may contain different
targets with the same name (or alias). Each target is still required
to be unique (i.e. they should all have different description strings).
This is mainly done to support katdal when opening observations with
targets that intentionally share the same name (e.g. when imaging a
calibrator, targets may differ only in tags).

The lookup dict changes from a mapping of string to Target to a mapping
of string to a list of Targets (i.e. all targets with the given name).

When looking up a target by name, return the most recently added target
with that name. The same goes for removing a target by name.

This implements JIRA ticket SR-1464.