Remove Catalogue.iternames method
Even though this is not underscored and therefore nominally part of the
public interface of Catalogue, it seems OK to remove without warning as
it was introduced as part of the tab completion machinery and never
used elsewhere (within SDP and CAM, at least).

The current version of the function is outdated as it will generate
duplicate names when targets with the same name goes into the Catalogue.
The lookup keys are a more direct way to get to searchable names.

If there is still a need for this in future, rather reintroduce it as
Catalogue.names, since the `iter` prefix has generally fallen out of
favour in Python 3+.