Improve UVW coordinate calculations
See the comments added to the code for detail about why this approach
improves things. As a side benefit, the special case handling for the
poles can be removed. This will change the results at the poles, in that
orientation will be determined by the given RA.

The golden RA/DEC in the unit tests changed. These are just a paste of
what is coming out, rather than independently verified results.