Add machinery to handle out-of-range inputs
The new `katpoint.projection.OutOfRange` class handles out-of-range
values in the spherical projection routines. Inspired by NumPy's seterr
machinery, it maintains a global treatment option that can be changed
permanently or temporarily using a with-statement. The following
treatment options are available:

  - 'raise': raise :class:`ProjectionInputsOutOfRangeError` (default)
  - 'nan': replace out-of-range values with NaNs
  - 'clip': replace out-of-range values with nearest valid values

The new exception is more specific but still derives from ValueError
to maintain backwards compatibility. The default treatment is 'raise'
for the same reason.

This addresses JIRA tickets SR-1502 and SR-1511.