Convert clip_2d to safe_scale
We don't use the full functionality of clipping 2D points to a torus.
Instead, we clip it to a circle to force a new radius. We also don't
change the default (x, y) to use when input vector has 0 length.

The clincher was the impending NaN support. If the new radius is NaN,
it would be nice to have the (x, y) output of `clip_2d` NaN too, but
`np.clip(x, np.nan, np.nan)` remains `x`. Once we get rid of `np.clip`,
we might as well turn `clip_2d` into what it really does: a safe way
to scale 2D vectors to have a new radius. And we gotta have a SAFe
function to scale well.