SPR1-900 Simplify angle conversion, close loophole
There is no need to support unit strings beyond 'd' and 'h', since
`_to_angle` is private and under our control. This simplifies things.

Ensure that `angle_from_degrees` returns an `ephem.degrees` object,
otherwise the description string won't round-trip. That's because the
angle object is expressed in sexagesimal form in its native format.
So an angle '4h' becomes an `ephem.hours` object with string
representation '4:00:00.00' but we want '60:00:00.00' if it is to be in
degrees. Add a test for this and do the same for `angle_from_hours`.