SPR1-3161 Restore description strings in NumPy 2.0
Since NumPy 2.0 the `repr` of NumPy data types like `np.int64` contains
the type. That is, `repr(np.int64(42))` now returns `"np.int64(42)"`
instead of the plain `"42"`. This breaks the flux model description
strings, where `repr` was previously introduced to preserve the
precision of the coefficients. The solution is to cast the `np.float64`
coefficients emerging from the ndarray to POD `float` before the `repr`.